+++ lib/tracker_web/controllers/user_controller.ex
@@ -7,8 +7,13 @@ defmodule TrackerWeb.UserController do
   action_fallback TrackerWeb.FallbackController
 
   def index(conn, _params) do
-    users = Users.list_users()
-    render(conn, "index.json", users: users)
+    if (Map.has_key?(Map.get(conn, :assigns), :authenticated_user_id)) do
+      users = Users.list_users()
+      render(conn, "index.json", users: users)
+    else
+      conn
+      |> send_resp(:unauthorized, "Must be logged in to access this resource")
+    end
   end
 
   def create(conn, %{"user" => user_params}) do
@@ -20,23 +25,53 @@ defmodule TrackerWeb.UserController do
     end
   end
 
+  def create(conn, %{"credentials" => user_credentials}) do
+    user = Users.authenticate(user_credentials["username"], user_credentials["password"])
+
+    if (user == nil) do
+      with {:unauthorized} do
+        conn
+        |>put_status(:unauthorized)
+        |>render("error.json", message: "Invalid credentials")
+      end
+    else
+      token = Phoenix.Token.sign(TrackerWeb.Endpoint, Application.get_env(:tracker, :app_salt), user.id)
+      render(conn, "authentication.json", user: user, token: token)
+    end
+  end
+
   def show(conn, %{"id" => id}) do
-    user = Users.get_user!(id)
-    render(conn, "show.json", user: user)
+    if (Map.has_key?(Map.get(conn, :assigns), :authenticated_user_id)) do
+      user = Users.get_user!(id)
+      render(conn, "show.json", user: user)
+    else
+      conn
+      |> send_resp(:unauthorized, "Must be logged in to access this resource")
+    end
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Users.get_user!(id)
+    if (Map.has_key?(Map.get(conn, :assigns), :authenticated_user_id)) do
+      user = Users.get_user!(id)
 
-    with {:ok, %User{} = user} <- Users.update_user(user, user_params) do
-      render(conn, "show.json", user: user)
+      with {:ok, %User{} = user} <- Users.update_user(user, user_params) do
+        render(conn, "show.json", user: user)
+      end
+    else
+      conn
+      |> send_resp(:unauthorized, "Must be logged in to access this resource")
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Users.get_user!(id)
-    with {:ok, %User{}} <- Users.delete_user(user) do
-      send_resp(conn, :no_content, "")
+    if (Map.has_key?(Map.get(conn, :assigns), :authenticated_user_id)) do
+      user = Users.get_user!(id)
+      with {:ok, %User{}} <- Users.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      conn
+      |> send_resp(:unauthorized, "Must be logged in to access this resource")
     end
   end
 end
